Fix next line adjustment for one-line regions
For some regions, we try to move the beginning of the region to the next line to
better match emitted debug info (where state changes like assignment appear as
taking effect on the next line). However, we sometime are working with a region
that only covers a single line (such as a macro invocation). Trying to move the
beginning of single-line region forward creates an invalid region, so we now
fallback to leaving the region as-is for such cases, which also seems to match
what is emitted.